+++ demo/demo/settings/production.py
@@ -1,5 +1,6 @@
+import os
+
 import sentry_sdk
-from decouple import config
 from sentry_sdk.integrations.django import DjangoIntegration
 
 from .base import *  # noqa
@@ -13,7 +14,7 @@ USE_X_FORWARDED_HOST = True
 SECURE_PROXY_SSL_HEADER = ("HTTP_X_FORWARDED_PROTO", "https")
 
 sentry_sdk.init(
-    dsn=config("SENTRY_DSN"),
+    dsn=os.environ["SENTRY_DSN"],
     integrations=[DjangoIntegration()],
     # Less transactions
     traces_sample_rate=0.01,