+++ demo/demo/settings/test.py
@@ -1,0 +1,3 @@
+from .development import *  # noqa
+
+INSTALLED_APPS.append("semantic_admin.tests.app.apps.SemanticAdminTestsAppConfig")  # noqa: F405