+++ demo/demo_app/factory.py
@@ -10,7 +10,7 @@ fake = Faker()
 
 class PersonFactory(DjangoModelFactory):
     name = LazyAttribute(lambda x: fake.first_name())
-    url = LazyAttribute(lambda x: fake.safe_domain_name())
+    url = LazyAttribute(lambda x: f"https://{fake.safe_domain_name()}")
     email = LazyAttribute(lambda x: fake.email())
     birthday = LazyAttribute(lambda x: fake.past_date())