+++ CHANGES.md
@@ -1,3 +1,42 @@
+v0.6.8
+------
+* Upgrade Fomantic v2.9.4 and use Django admin jQuery
+* `SemanticTabularInline` delete checkbox fix
+* Deduplicate JavaScript catalog
+
+v0.6.7
+------
+* Require Django >= 4.2
+* Update templates and JS
+* Fix `format_html()` deprecations for Django 6
+
+v0.6.6
+------
+* Improve demo login page more
+* Change/view permission fixes
+
+v0.6.5
+------
+* Move django-import-export customizations to `semantic_admin.contrib.import_export`
+
+v0.6.4
+------
+* Improve demo login page
+
+v0.6.3
+------
+* Fix `SplitDateTimeField` handling
+* Use sixteen wide field for header
+
+v0.6.2
+------
+* Add a `.hidden` CSS class for admin
+
+v0.6.1
+------
+* Add `jsi18n` support to the change list
+* Improve static handling and formfield overrides
+
 v0.6.0
 ------
 * Require Python >= 3.10
@@ -5,4 +44,4 @@ v0.6.0
 
 v0.5.0
 ------
-* Forms were moved to [django-semantic-forms](https://github.com/globophobe/django-semantic-forms) so that they may be used outside of the admin.
+* Forms were moved to [django-semantic-forms](https://github.com/globophobe/django-semantic-forms) so that they may be used outside of the admin